Upgrades the Plugin by moving into the architecture & optimizing (PR #2656)
* Moves Plugin file into `inc` and renames file.
* Moves closure to a method for getting the container.
* Injects Container instance into the Plugin.
* Refactors to distribute tasks.
* Optimization: stores rocket_valid_key() into a property for reuse.
* Create options instance and store on object instead of calling get_rocket_option().
* Moves the event manager to a property & reuse instead of pulling out of the container.
* Removes the file direct call guard clause.
